split out scan modules
