+++ src/scan/scanner/error.ts
@@ -1,0 +1,1 @@
+export class BarcodeDetectionError extends Error {}
