+++ src/scan/scanner/type.ts
@@ -1,0 +1,5 @@
+export type BarcodeScanner = {
+  scanBlob: (blob: Blob) => Promise<string>;
+  scanUrl: (url: string) => Promise<string>;
+  scanVideo: (video: HTMLVideoElement) => Promise<string>;
+};
